pyproject.toml
[tool.poetry]
name = "crud-backend"
version = "0.1.0"
description = ""
authors = ["AOBM <oblancomorales@gmail.com>"]
readme = "README.md"
packages = [{include = "crud_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
psycopg2-binary = "^2.9.9"
django-configurations = "^2.5.1"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.7"
requests = "2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
